requirements.txt
# requirements.txt

openai
pandas
requests
beautifulsoup4
streamlit
supabase
python-dotenv
plotly